Update 20210831_biomassUsingFAOSTAT.R
Add more details on the methods and data sources. Please note that the biomass methods on the dashboard are two: the "GBADs "one should be the method using the slaughterweight.  The other option named "WOAH" should be from the spreadsheet model of Morgan, which was not included in this R code. I think the Informatic theme worked with Morgan on that method.